new features frontend

+++ backend/src/backend/controllers/links_controller.ts
@@ -16,24 +16,48 @@ export const linkSchema = {
     field: joi.string()
 };
 
-export function getHandlers(linkRepo: Repository <Link>) {
+interface LinkPreviewDetails {
+    id: number;
+    userId: number;
+    email: string;
+    name: string;
+    title: string;
+    question: string;
+    field: string;
+    date: string;
+    replyCount: number | null;
+}
+
+export function getHandlers(linkRepo: Repository<Link>) {
 
     const getAllLinks = (req: express.Request, res: express.Response) => {
         (async () => {
             try {
-                const links = await linkRepo.createQueryBuilder("link")
-                    .leftJoinAndSelect("link.user", "user")
-                    .leftJoinAndSelect("link.reply", "reply")
-                    .getMany();
+                const queryResult: LinkPreviewDetails[] = await linkRepo.query(`
+                    SELECT
+                        "link"."id",
+                        "link"."userId",
+                        "user"."email",
+                        "user"."name",
+                        "link"."title",
+                        "link"."question",
+                        "link"."field",
+                        "link"."date",
+                        count("reply"."id") "replyCount"
+                    FROM "link" "link"
+                    LEFT JOIN "user" "user" ON "user"."id" = "link"."userId"
+                    LEFT JOIN "reply" "reply" ON "reply"."linkId" = "link"."id"
+                    GROUP BY "link"."id", "user"."name", "user"."email"
+                `);
 
-                // const links = await linkRepo.find();
-                res.json(links);
-            }catch (err) {
+                res.json(queryResult);
+
+            } catch (err) {
                 // Handle unexpected errors
                 console.error(err);
                 res.status(500)
-                   .json({ error: "Internal server error"})
-                   .send();
+                    .json({ error: "Internal server error" })
+                    .send();
             }
         })();
     }
@@ -46,19 +70,61 @@ export function getHandlers(linkRepo: Repository <Link>) {
                 const idStr = req.params.id;
                 const linkId = { id: parseInt(idStr) };
                 const idValidationresult = joi.validate(linkId, linkIdSchema);
-                
+
                 if (idValidationresult.error) {
                     res.status(400).json({ error: "Bad request" }).send();
-                } {
-                    const link = await linkRepo.createQueryBuilder("link")
-                                                .leftJoinAndSelect("link.reply", "reply")
-                                                .where("link.id = :id", { id: linkId.id })
-                                                .getOne();
-                    if (link === undefined) {
+                } else {
+
+                    const queryResult = await linkRepo.query(`
+                        SELECT
+                            "link"."id",
+                            "link"."userId",
+                            "user"."email",
+                            "link"."title",
+                            "link"."question",
+                            "link"."date",
+                            count("reply"."id") "replyCount"
+                        FROM "link" "link"
+                        LEFT JOIN "user" "user" ON "user"."id" = "link"."userId"
+                        LEFT JOIN "reply" "reply" ON "reply"."linkId" = "link"."id"
+                        WHERE "link"."id" = $1
+                        GROUP BY "link"."id", "user"."name", "user"."email"
+                    `, [ linkId.id ]);
+
+                    const linkPreviewDetails: LinkPreviewDetails = queryResult[0];
+
+                    if (linkPreviewDetails === undefined) {
                         res.status(404)
-                           .json({ error: "Not found"})
-                           .send();
+                            .json({ error: "Not found" })
+                            .send();
                     } else {
+
+                        const replies = await linkRepo.query(`
+                            SELECT
+                                "reply"."id",
+                                "reply"."userId",
+                                "user"."name",
+                                "user"."email",
+                                "reply","linkId",
+                                "reply"."text",
+                                "reply"."date"
+                            FROM "reply" "reply"
+                            LEFT JOIN "user" "user" ON "user"."id" = "reply"."userId"
+                            WHERE "linkId" = $1
+                        `, [linkId.id]);
+
+                        const link = {
+                            id: linkPreviewDetails.id,
+                            userId: linkPreviewDetails.userId,
+                            name: linkPreviewDetails.name,
+                            email: linkPreviewDetails.email,
+                            title: linkPreviewDetails.title,
+                            question: linkPreviewDetails.question,
+                            date: linkPreviewDetails.date,
+                            replyCount: linkPreviewDetails.replyCount,
+                            replies: replies
+                        };
+
                         res.json(link);
                     }
                 }
@@ -66,8 +132,8 @@ export function getHandlers(linkRepo: Repository <Link>) {
                 // Handle unexpected errors
                 console.error(err);
                 res.status(500)
-                   .json({ error: "Internal server error"})
-                   .send();
+                    .json({ error: "Internal server error" })
+                    .send();
             }
         })();
     }
@@ -75,7 +141,7 @@ export function getHandlers(linkRepo: Repository <Link>) {
     const createLink = (req: express.Request, res: express.Response) => {
         (async () => {
             try {
-                
+
                 // The request userId property is set by the authMiddleware
                 // if it is undefined it means that we forgot the authMiddleware
                 if ((req as AuthenticatedRequest).userId === undefined) {
@@ -87,17 +153,18 @@ export function getHandlers(linkRepo: Repository <Link>) {
                 const result = joi.validate(newLink, linkSchema);
 
                 if (result.error) {
-                    res.json({ msg: `Invalid user details in body!`}).status(400).send();
+                    res.json({ msg: `Invalid user details in body!` }).status(400).send();
                 } else {
 
                     // Create new link
                     const linkToBeSaved = new Link();
                     linkToBeSaved.userId = (req as AuthenticatedRequest).userId;
-                    linkToBeSaved.question = req.body.question;
+                    linkToBeSaved.question = newLink.question;
                     linkToBeSaved.title = newLink.title;
                     linkToBeSaved.field = newLink.field;
-                    let today = new Date();
-                    linkToBeSaved.date = "at " + today.getDate() + "-" +(today.getMonth() +1) + "-" + today.getFullYear() + " " + today.getHours() + ":" + today.getMinutes(); 
+                    linkToBeSaved.date = new Date();
+                    // let today = new Date();
+                    // linkToBeSaved.date = "at " + today.getDate() + "-" +(today.getMonth() +1) + "-" + today.getFullYear() + " " + today.getHours() + ":" + today.getMinutes(); 
                     const savedLink = await linkRepo.save(linkToBeSaved);
                     res.json(savedLink).send();
                 }
@@ -106,8 +173,8 @@ export function getHandlers(linkRepo: Repository <Link>) {
                 // Handle unexpected errors
                 console.error(err);
                 res.status(500)
-                   .json({ error: "Internal server error"})
-                   .send();
+                    .json({ error: "Internal server error" })
+                    .send();
             }
         })();
     }
@@ -126,7 +193,7 @@ export function getHandlers(linkRepo: Repository <Link>) {
                 const idStr = req.params.id;
                 const linkId = { id: parseInt(idStr) };
                 const idValidationresult = joi.validate(linkId, linkIdSchema);
-                
+
                 if (idValidationresult.error) {
                     res.status(400).json({ error: "Bad request" }).send();
                 } else {
@@ -137,25 +204,34 @@ export function getHandlers(linkRepo: Repository <Link>) {
                     // If link not found return 404 not found
                     if (link === undefined) {
                         res.status(404)
-                           .json({ error: "Not found"})
-                           .send();
+                            .json({ error: "Not found" })
+                            .send();
                     } else {
 
-                        // If lik was found, remove it from DB
-                        await linkRepo.remove(link);
-                        res.json({ msg: "OK" }).send();
+                        // If link was found, remove it from DB
+                        const userId = (req as AuthenticatedRequest).userId;
+                        const ownerId = link.userId;
+
+                        if (userId !== ownerId) {
+                            res.status(403)
+                                .json({ msg: `The current user is not the author of the link` })
+                                .send();
+                        } else {
+                            await linkRepo.remove(link);
+                            res.json({ msg: "OK" }).send();
+                        }
                     }
                 }
             } catch (err) {
                 // Handle unexpected errors
                 console.error(err);
                 res.status(500)
-                   .json({ error: "Internal server error"})
-                   .send();
+                    .json({ error: "Internal server error" })
+                    .send();
             }
         })();
     }
-  
+
     return {
         getAllLinks,
         getLinkById,
